Renamed research_agents folder
