+++ .vscode/settings.json
@@ -24,10 +24,7 @@
     "editor.defaultFormatter": "ms-python.black-formatter",
     "editor.formatOnSave": true,
   },
-  "isort.args": [
-    "--profile",
-    "black"
-  ],
+  "isort.args":["--profile", "black","--line-length", "79"],
   "jupyter.debugJustMyCode": true,
   "debugpy.debugJustMyCode": true,
 }
