+++ code_tests/expensive_tests__run_individually/test_ex_base_rate_responder.py
@@ -1,7 +1,7 @@
 import asyncio
 from datetime import datetime
 
-from forecasting_tools.research_agents.base_rate_researcher import (
+from forecasting_tools.agents_and_tools.base_rate_researcher import (
     BaseRateReport,
     BaseRateResearcher,
     DenominatorOption,
