+++ code_tests/expensive_tests__run_individually/test_ex_key_factors.py
@@ -2,14 +2,14 @@ import logging
 
 import pytest
 
+from forecasting_tools.agents_and_tools.key_factors_researcher import (
+    KeyFactorsResearcher,
+    ScoredKeyFactor,
+)
 from forecasting_tools.ai_models.resource_managers.monetary_cost_manager import (
     MonetaryCostManager,
 )
 from forecasting_tools.forecast_helpers.metaculus_api import MetaculusApi
-from forecasting_tools.research_agents.key_factors_researcher import (
-    KeyFactorsResearcher,
-    ScoredKeyFactor,
-)
 
 logger = logging.getLogger(__name__)
