+++ code_tests/expensive_tests__run_individually/test_ex_niche_list_researcher.py
@@ -3,14 +3,14 @@ import logging
 
 import pytest
 
+from forecasting_tools.agents_and_tools.niche_list_researcher import (
+    FactCheckedItem,
+    NicheListResearcher,
+)
 from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
 from forecasting_tools.ai_models.deprecated_model_classes.gpt4o import Gpt4o
 from forecasting_tools.ai_models.resource_managers.monetary_cost_manager import (
     MonetaryCostManager,
-)
-from forecasting_tools.research_agents.niche_list_researcher import (
-    FactCheckedItem,
-    NicheListResearcher,
 )
 
 logger = logging.getLogger(__name__)
