+++ code_tests/low_cost_or_live_api_tests/test_base_rate_responder.py
@@ -1,6 +1,6 @@
 import pytest
 
-from forecasting_tools.research_agents.base_rate_researcher import (
+from forecasting_tools.agents_and_tools.base_rate_researcher import (
     BaseRateResearcher,
 )
