+++ forecasting_tools/agents_and_tools/estimator.py
@@ -1,11 +1,11 @@
 from datetime import datetime
 from typing import NamedTuple
 
-from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
-from forecasting_tools.forecast_helpers.smart_searcher import SmartSearcher
-from forecasting_tools.research_agents.general_researcher import (
+from forecasting_tools.agents_and_tools.general_researcher import (
     GeneralResearcher,
 )
+from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
+from forecasting_tools.forecast_helpers.smart_searcher import SmartSearcher
 
 
 class EstimationResult(NamedTuple):
