+++ forecasting_tools/agents_and_tools/research_coordinator.py
@@ -2,23 +2,23 @@ from __future__ import annotations
 
 import logging
 
+from forecasting_tools.agents_and_tools.base_rate_researcher import (
+    BaseRateReport,
+    BaseRateResearcher,
+)
+from forecasting_tools.agents_and_tools.general_researcher import (
+    GeneralResearcher,
+)
+from forecasting_tools.agents_and_tools.question_responder import (
+    QuestionResponder,
+)
+from forecasting_tools.agents_and_tools.question_router import QuestionRouter
 from forecasting_tools.ai_models.ai_utils.ai_misc import (
     clean_indents,
     strip_code_block_markdown,
 )
 from forecasting_tools.ai_models.configured_llms import BasicLlm
 from forecasting_tools.data_models.questions import MetaculusQuestion
-from forecasting_tools.research_agents.base_rate_researcher import (
-    BaseRateReport,
-    BaseRateResearcher,
-)
-from forecasting_tools.research_agents.general_researcher import (
-    GeneralResearcher,
-)
-from forecasting_tools.research_agents.question_responder import (
-    QuestionResponder,
-)
-from forecasting_tools.research_agents.question_router import QuestionRouter
 from forecasting_tools.util import async_batching
 
 logger = logging.getLogger(__name__)
