+++ forecasting_tools/front_end/app_pages/estimator_page.py
@@ -5,6 +5,7 @@ import logging
 import streamlit as st
 from pydantic import BaseModel
 
+from forecasting_tools.agents_and_tools.estimator import Estimator
 from forecasting_tools.ai_models.resource_managers.monetary_cost_manager import (
     MonetaryCostManager,
 )
@@ -13,7 +14,6 @@ from forecasting_tools.forecast_helpers.forecast_database_manager import (
     ForecastRunType,
 )
 from forecasting_tools.front_end.helpers.tool_page import ToolPage
-from forecasting_tools.research_agents.estimator import Estimator
 from forecasting_tools.util.jsonable import Jsonable
 
 logger = logging.getLogger(__name__)
