+++ forecasting_tools/front_end/app_pages/key_factors_page.py
@@ -14,6 +14,10 @@ current_dir = os.path.dirname(os.path.abspath(__file__))
 top_level_dir = os.path.abspath(os.path.join(current_dir, "../../../"))
 sys.path.append(top_level_dir)
 
+from forecasting_tools.agents_and_tools.key_factors_researcher import (
+    KeyFactorsResearcher,
+    ScoredKeyFactor,
+)
 from forecasting_tools.ai_models.resource_managers.monetary_cost_manager import (
     MonetaryCostManager,
 )
@@ -26,10 +30,6 @@ from forecasting_tools.front_end.helpers.report_displayer import (
     ReportDisplayer,
 )
 from forecasting_tools.front_end.helpers.tool_page import ToolPage
-from forecasting_tools.research_agents.key_factors_researcher import (
-    KeyFactorsResearcher,
-    ScoredKeyFactor,
-)
 from forecasting_tools.util.jsonable import Jsonable
 
 logger = logging.getLogger(__name__)
