+++ forecasting_tools/front_end/app_pages/question_generation_page.py
@@ -7,6 +7,11 @@ from datetime import datetime, timedelta
 import streamlit as st
 from pydantic import BaseModel
 
+from forecasting_tools.agents_and_tools.question_generator import (
+    GeneratedQuestion,
+    QuestionGenerator,
+    TopicGenerator,
+)
 from forecasting_tools.ai_models.general_llm import GeneralLlm
 from forecasting_tools.ai_models.resource_managers.monetary_cost_manager import (
     MonetaryCostManager,
@@ -20,11 +25,6 @@ from forecasting_tools.front_end.helpers.report_displayer import (
     ReportDisplayer,
 )
 from forecasting_tools.front_end.helpers.tool_page import ToolPage
-from forecasting_tools.research_agents.question_generator import (
-    GeneratedQuestion,
-    QuestionGenerator,
-    TopicGenerator,
-)
 from forecasting_tools.util.jsonable import Jsonable
 
 logger = logging.getLogger(__name__)
