+++ pyproject.toml
@@ -43,6 +43,7 @@ plotly = ">=5.24.1,<7.0.0"
 pandas = "^2.2.3"
 streamlit = "^1.20.0"
 asknews = ">=0.9.1,<0.12.0"
+openai-agents = {extras = ["litellm"], version = "^0.0.14"}
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^4.0.1"
